feat(sleep): support fractional sleep durations and enhance validation logic

+++ automation-service/internal/engine/definition.go
@@ -124,7 +124,7 @@ type NotifyEmailRecipient struct {
 }
 
 type LogicSleep struct {
-	DurationSec int `json:"duration_sec"`
+	DurationSec float64 `json:"duration_sec"`
 }
 
 type LogicIf struct {
